jobs,catpb,descpb: move the definition of JobID down to catpb, forward up
This allows the JobID to be used in descriptors. Adds a little bit of type
safety here and there.

Release note: None

+++ pkg/sql/indexbackfiller_test.go
@@ -498,7 +498,7 @@ INSERT INTO foo VALUES (1), (10), (100);
 				return err
 			}
 			mut.MutationJobs = append(mut.MutationJobs, descpb.TableDescriptor_MutationJob{
-				JobID:      int64(jobID),
+				JobID:      jobID,
 				MutationID: 1,
 			})
 			jobToBlock.Store(jobID)
